Move webhook scoping rules into a predicates directory
