+++ staging/src/k8s.io/apiserver/pkg/admission/plugin/webhook/predicates/namespace/doc.go
